Separate into two files the tests for Passcode and Biometrical inside a folder caled Security
